style(*): optimizes code styles
